refactor: move ComponentData and ComponentType to their own file
