+++ lib/logic/component_service.dart
@@ -1,5 +1,7 @@
 import 'package:editicert/logic/component_index_service.dart';
 import 'package:editicert/models/component.dart';
+import 'package:editicert/models/component_data.dart';
+import 'package:editicert/models/component_type.dart';
 import 'package:editicert/util/utils.dart';
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
@@ -81,74 +83,4 @@ class ComponentService {
       _state = [..._state..removeAt(selectedIndex)];
     }
   }
-}
-
-@immutable
-class ComponentData {
-  const ComponentData({
-    this.name = 'Component',
-    this.component = const Component(
-      Offset.zero,
-      Size(100, 100),
-      0,
-      false,
-      false,
-    ),
-    this.color = const Color(0xFF9E9E9E),
-    this.borderRadius = BorderRadius.zero,
-    this.border = const Border(),
-    this.shadow = const [],
-    this.content = const SizedBox.shrink(),
-    this.hidden = false,
-    this.locked = false,
-    this.type = ComponentType.frame,
-    this.textController,
-  });
-
-  final String name;
-  final Component component;
-  final Color color;
-  final BorderRadius borderRadius;
-  final Border border;
-  final List<BoxShadow> shadow;
-  final Widget content;
-  final bool hidden;
-  final bool locked;
-  final ComponentType type;
-  final TextEditingController? textController;
-
-  ComponentData copyWith({
-    required String? name,
-    required Component? component,
-    required Color? color,
-    required BorderRadius? borderRadius,
-    required Border? border,
-    required List<BoxShadow>? shadow,
-    required Widget? content,
-    required bool? hidden,
-    required bool? locked,
-    required ComponentType? type,
-    required TextEditingController? textController,
-  }) =>
-      ComponentData(
-        name: name ?? this.name,
-        component: component ?? this.component,
-        color: color ?? this.color,
-        borderRadius: borderRadius ?? this.borderRadius,
-        border: border ?? this.border,
-        shadow: shadow ?? this.shadow,
-        content: content ?? this.content,
-        hidden: hidden ?? this.hidden,
-        locked: locked ?? this.locked,
-        type: type ?? this.type,
-        textController: textController ?? this.textController,
-      );
-}
-
-enum ComponentType {
-  frame,
-  text,
-  image,
-  rectangle,
-  other,
 }
