+++ lib/models/component_type.dart
@@ -1,0 +1,7 @@
+enum ComponentType {
+  frame,
+  text,
+  image,
+  rectangle,
+  other,
+}
